services/rag-service/pyproject.toml
[tool.poetry]
name = "rag-service"
version = "1.0.0"
description = "RAG服务 - 苏柯生命平台的知识增强检索生成服务"
authors = ["Suoke Life Team <team@suokelife.com>"]
readme = "README.md"
packages = [{include = "services"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
aiohttp = "^3.8.5"
openai = "^1.3.0"
pydantic = "^2.3.0"
python-dotenv = "^1.0.0"
redis = "^5.0.0"
qdrant-client = "^1.6.0"
grpcio = "^1.59.0"
grpcio-tools = "^1.59.0"
loguru = "^0.7.0"
tenacity = "^8.2.3"
pyyaml = "^6.0.1"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
opentelemetry-exporter-otlp = "^1.20.0"
opentelemetry-instrumentation = "^0.40b0"
opentelemetry-instrumentation-aiohttp = "^0.40b0"
opentelemetry-instrumentation-grpc = "^0.40b0"
opentelemetry-instrumentation-redis = "^0.40b0"
prometheus-client = "^0.17.1"
motor = "^3.3.1"
nltk = "^3.8.1"
tiktoken = "^0.5.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
types-pyyaml = "^6.0.12.12"
httpx = "^0.24.1"
pytest-mock = "^3.11.1"
faker = "^19.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto" 